cookbook/python-batch-flow/requirements.txt
brainyflow
numpy
Pillow>=10.0.0
